GoogleGenAI.GoogleProvider

GoogleGenAI.count_tokens

GoogleGenAI.embed_content

GoogleGenAI.generate_content

GoogleGenAI.list_models

API

GoogleGenAI.GoogleProvider — Type

Base.@kwdef struct GoogleProvider <: AbstractGoogleProvider
    api_key::String = ""
    base_url::String = "https://generativelanguage.googleapis.com"
    api_version::String = "v1beta"
end

A configuration object used to set up and authenticate requests to the Google Generative Language API.

Fields

api_key::String: Your Google API key.

base_url::String: The base URL for the Google Generative Language API. The default is set to "https://generativelanguage.googleapis.com".

api_version::String: The version of the API you wish to access. The default is set to "v1beta".

GoogleGenAI.generate_content — Function

generate_content(provider::AbstractGoogleProvider, model_name::String, prompt::String, image_path::String; api_kwargs=NamedTuple(), https_kwargs=NamedTuple()) -> NamedTuple
generate_content(api_key::String, model_name::String, prompt::String, image_path::String; api_kwargs=NamedTuple(), https_kwargs=NamedTuple()) -> NamedTuple

generate_content(provider::AbstractGoogleProvider, model_name::String, conversation::Vector{Dict{Symbol,Any}}; api_kwargs=NamedTuple(), https_kwargs=NamedTuple()) -> NamedTuple
generate_content(api_key::String, model_name::String, conversation::Vector{Dict{Symbol,Any}}; api_kwargs=NamedTuple(), https_kwargs=NamedTuple()) -> NamedTuple

Generate content based on a combination of text prompt and an image (optional).

Arguments

provider::AbstractGoogleProvider: The provider instance for API requests.

api_key::String: Your Google API key as a string.

model_name::String: The model to use for content generation.

prompt::String: The text prompt to accompany the image.

image_path::String (optional): The path to the image file to include in the request.

API Keyword Arguments

temperature::Float64 (optional): Controls the randomness in the generation process. Higher values result in more random outputs. Typically ranges between 0 and 1.

candidate_count::Int (optional): The number of generation candidates to consider. Currently, only one candidate can be specified.

max_output_tokens::Int (optional): The maximum number of tokens that the generated content should contain.

stop_sequences::Vector{String} (optional): A list of sequences where the generation should stop. Useful for defining natural endpoints in generated content.

safety_settings::Vector{Dict} (optional): Settings to control the safety aspects of the generated content, such as filtering out unsafe or inappropriate content.

HTTP Kwargs

All keyword arguments supported by the HTTP.request function. Documentation can be found here: https://juliaweb.github.io/HTTP.jl/stable/reference/#HTTP.request.

Returns

NamedTuple: A named tuple containing the following keys:

candidates: A vector of dictionaries, each representing a generation candidate.

safety_ratings: A dictionary containing safety ratings for the prompt feedback.

text: A string representing the concatenated text from all candidates.

response_status: An integer representing the HTTP response status code.

finish_reason: A string indicating the reason why the generation process was finished.

GoogleGenAI.count_tokens — Function

count_tokens(provider::AbstractGoogleProvider, model_name::String, prompt::String) -> Int
count_tokens(api_key::String, model_name::String, prompt::String) -> Int

Calculate the number of tokens generated by the specified model for a given prompt string.

Arguments

provider::AbstractGoogleProvider: The provider instance containing API key and base URL information.

api_key::String: Your Google API key as a string.

model_name::String: The name of the model to use for generating content.

prompt::String: The prompt prompt based on which the text is generated.

Returns

Int: The total number of tokens that the given prompt string would be broken into by the specified model's tokenizer.

GoogleGenAI.embed_content — Function

embed_content(provider::AbstractGoogleProvider, model_name::String, prompt::String https_kwargs=NamedTuple()) -> NamedTuple
embed_content(api_key::String, model_name::String, prompt::String https_kwargs=NamedTuple()) -> NamedTuple
embed_content(provider::AbstractGoogleProvider, model_name::String, prompts::Vector{String} https_kwargs=NamedTuple()) -> Vector{NamedTuple}
embed_content(api_key::String, model_name::String, prompts::Vector{String}, https_kwargs=NamedTuple()) -> Vector{NamedTuple}

Generate an embedding for the given prompt text using the specified model.

Arguments

provider::AbstractGoogleProvider: The provider instance containing API key and base URL information.

api_key::String: Your Google API key as a string.

model_name::String: The name of the model to use for generating content.

prompt::String: The prompt prompt based on which the text is generated.

HTTP Kwargs

All keyword arguments supported by the HTTP.request function. Documentation can be found here: https://juliaweb.github.io/HTTP.jl/stable/reference/#HTTP.request.

Returns

NamedTuple: A named tuple containing the following keys:

values: A vector of Float64 representing the embedding values for the given prompt.

response_status: An integer representing the HTTP response status code.

GoogleGenAI.list_models — Function

list_models(provider::AbstractGoogleProvider) -> Vector{Dict}
list_models(api_key::String) -> Vector{Dict}

Retrieve a list of available models along with their details from the Google AI API.

Arguments

provider::AbstractGoogleProvider: The provider instance containing API key and base URL information.

api_key::String: Your Google API key as a string.

Returns

Vector{Dict}: A list of dictionaries, each containing details about an available model.